fix(eng-360): drop redundant int64(st.Bsize) cast (unconvert lint)

+++ internal/backend/docker/retention_accounting.go
@@ -44,7 +44,7 @@ func (b *Backend) warnIfOverProvisioned() {
 		b.logger.Warn("capacity check: statfs failed", "path", b.cfg.VolumeDataPath, "error", err)
 		return
 	}
-	fsTotalMB := int64(st.Blocks) * int64(st.Bsize) / bytesPerMiB
+	fsTotalMB := int64(st.Blocks) * st.Bsize / bytesPerMiB
 	if diskOverProvisioned(b.cfg.TotalDiskMB, fsTotalMB) {
 		b.logger.Warn("total_disk_mb exceeds the data filesystem's total capacity: over-commit risk (retained+live volumes can exhaust physical disk → tenant ENOSPC). Size total_disk_mb at or below usable capacity, leaving headroom for root-reserved blocks and non-fred consumers.",
 			"total_disk_mb", b.cfg.TotalDiskMB, "fs_total_mb", fsTotalMB, "path", b.cfg.VolumeDataPath)
